[이나라][Player] 몬스터 공격과 플레이어 피격 연동

+++ Assets/02.Scripts/Monster/IMonsterCombatable.cs
@@ -5,7 +5,7 @@ public interface IMonsterCombatable
     bool IsAttacking { get; }
     float AttackRange { get; }
 
-    float AttackPower { get; }
+    int AttackPower { get; }
 
     event Action OnAttackStarted;
     event Action OnAttackEnded;
